refactor(dashboards): extract computeMomPct to shared utils
Move computeMomPct to @lfx-one/shared/utils alongside the other trend
helpers. Remove local copies from email-tab and performance-marketing-tab.
Also cache .at() values in email-tab open-rate calculation to eliminate
non-null assertions.

LFXV2-1644
